Combine test classes into UnitTests and IntegrationTests
