+++ src/main/java/com/napier/sem/App.java
@@ -41,15 +41,19 @@ public class App
 
         UserPrompt prompt = new UserPrompt(dao);
 
-        try
+        if (connection != null)
         {
-            prompt.start();
+            try
+            {
+                prompt.start();
+            }
+            catch (NoSuchElementException e)
+            {
+                System.out.println(App.HORIZONTAL_LINE + "\nNo console detected, running example query");
+                for (Record record: dao.topNCountriesIn(App.REGION, "Caribbean", 5)) System.out.println(record);
+            }
         }
-        catch (NoSuchElementException e)
-        {
-            System.out.println(App.HORIZONTAL_LINE + "\nNo console detected, running example query");
-            for (Record record: dao.topNCountriesIn(App.REGION, "Caribbean", 5)) System.out.println(record);
-        }
+        else System.out.println("Connection was null");
 
         // Disconnect from database
         disconnect(connection);
