--- src/test/java/com/napier/sem/DAO_UnitTests.java
@@ -1,751 +1,0 @@
-package com.napier.sem;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.BeforeAll;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-public class DAO_UnitTests
-{
-    static DAO dao;
-
-    @BeforeAll
-    static void init() { dao = new DAO(null); }
-
-    /**
-     * Unit tests covering the DAO.getWhereCondition method
-     */
-    // test the whereCondition is null when the areaFilter is unexpected
-    @Test
-    public void getWhereCondition_unknownAreaFilterReturnsNull()
-    {
-        // given
-        String areaFilter = "planet";
-
-        // when
-        String whereCondition = DAO.getWhereCondition(areaFilter, "Earth");
-
-        // then
-        assertNull(whereCondition);
-    }
-
-    // test that the whereCondition contains our given area name
-    @Test
-    public void getWhereCondition_areaNameAppearsInWhereCondition()
-    {
-        // given
-        String areaName = "New York";
-
-        // when
-        String whereCondition = DAO.getWhereCondition(App.DISTRICT, areaName);
-
-        // then
-        assertTrue(whereCondition.contains(areaName));
-    }
-
-    // test that the whereCondition is null when areaName is null
-    @Test
-    public void getWhereCondition_nullAreaNameReturnsNull()
-    {
-        // when
-        String whereCondition = DAO.getWhereCondition(App.WORLD, null);
-
-        // then
-        assertNull(whereCondition);
-    }
-
-    // test that the whereCondition is null when areaFilter is null
-    @Test
-    public void getWhereCondition_nullAreaFilterReturnsNull()
-    {
-        // when
-        String whereCondition = DAO.getWhereCondition(null, "Earth");
-
-        // then
-        assertNull(whereCondition);
-    }
-
-    /**
-     * Unit tests covering the DAO.queryInvalid method
-     */
-    // test that a null whereCondition will result in queryInvalid being true
-    @Test
-    public void queryInvalid_falseWhenWhereConditionIsNull()
-    {
-        // when
-        boolean result = dao.queryInvalid("query",
-                null,
-                App.WORLD,
-                new ArrayList<>(),
-                1);
-
-        assertTrue(result);
-    }
-
-    // test that an n value of less than 1 will result in queryInvalid being true
-    @Test
-    public void queryInvalid_falseWhenNLessThanOne()
-    {
-        // given
-        int n = 0;
-
-        // when
-        boolean result = dao.queryInvalid("query",
-                "country.code LIKE '%'",
-                App.WORLD,
-                new ArrayList<>(),
-                n);
-
-        assertTrue(result);
-    }
-
-    // test that an area filter not in the valid area filters will result in queryInvalid being true
-    @Test
-    public void queryInvalid_falseWhenAreaFilterNotAllowed()
-    {
-        // given
-        ArrayList<String> validAreaFilters = new ArrayList<>(Arrays.asList(App.WORLD, App.CONTINENT, App.REGION));
-        String areaFilter = App.COUNTRY;
-
-        // when
-        boolean result = dao.queryInvalid("query",
-                "country.name LIKE '%'",
-                areaFilter,
-                validAreaFilters,
-                1);
-
-        assertTrue(result);
-    }
-
-    // test that queryInvalid is false when it is passed good values
-    @Test
-    public void queryInvalid_happyPath()
-    {
-        // given
-        String queryName = "query";
-        String whereCondition = "country.name LIKE '%'";
-        ArrayList<String> validAreaFilters = new ArrayList<>(Arrays.asList(App.WORLD, App.CONTINENT, App.REGION));
-        String areaFilter = App.WORLD;
-        int n = 5;
-
-        // when
-        boolean result = dao.queryInvalid(queryName,
-                whereCondition,
-                areaFilter,
-                validAreaFilters,
-                n);
-
-        assertFalse(result);
-    }
-
-    /**
-     * Unit tests covering the DAO.allCountriesIn method
-     */
-    // test that countries list is empty when areaFilter and areaName are null
-    @Test
-    public void allCountriesIn_bothArgumentsNullListEmpty()
-    {
-        // when
-        ArrayList<Record> countries = dao.allCountriesIn(null, null);
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // test that countries list is empty when areaFilter is null
-    @Test
-    public void allCountriesIn_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> countries = dao.allCountriesIn(areaFilter, "Scotland");
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // test that countries list is empty when areaName is null
-    @Test
-    public void allCountriesIn_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> countries = dao.allCountriesIn(App.COUNTRY, areaName);
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // check countries list is empty if called with an invalid areaFilter
-    @Test
-    public void allCountriesIn_areaFilterWrongListEmpty()
-    {
-        // given
-        String areaFilter = App.CITY;
-
-        // when
-        ArrayList<Record> countries = dao.allCountriesIn(areaFilter, "Glasgow");
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void allCountriesIn_happyPath()
-    {
-        // given
-        String areaFilter = App.COUNTRY;
-        String areaName = "France";
-
-        // then
-        dao.allCountriesIn(areaFilter, areaName); // No Error
-    }
-
-    /**
-     * Unit tests covering the DAO.topNCountriesIn method
-     */
-    // test that countries list is empty when areaFilter and areaName are null
-    @Test
-    public void topNCountriesIn_bothArgumentsNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-        String areaName = null;
-
-        // when
-        ArrayList<Record> countries = dao.topNCountriesIn(areaFilter, areaName, 5);
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // test that countries list is empty when areaFilter is null
-    @Test
-    public void topNCountriesIn_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> countries = dao.topNCountriesIn(areaFilter, "Scotland", 10);
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // test that countries list is empty when areaName is null
-    @Test
-    public void topNCountriesIn_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> countries = dao.topNCountriesIn(App.COUNTRY, areaName, 3);
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // test that n cannot be less than or equal to 0
-    @Test
-    public void topNCountriesIn_negativeNListEmpty()
-    {
-        // given
-        Integer n = -3;
-
-        // when
-        ArrayList<Record> countries = dao.topNCountriesIn(App.COUNTRY, "China", n);
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // check countries list is empty if called with an invalid areaFilter
-    @Test
-    public void topNCountriesIn_areaFilterWrongListEmpty() {
-        // given
-        String areaFilter = App.DISTRICT;
-
-        // when
-        ArrayList<Record> countries = dao.topNCountriesIn(areaFilter, "Scotland", 15);
-
-        // then
-        assertEquals(0, countries.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void topNCountriesIn_happyPath()
-    {
-        // given
-        String areaFilter = App.COUNTRY;
-        String areaName = "France";
-
-        // then
-        dao.topNCountriesIn(areaFilter, areaName, 15); // No Error
-    }
-
-    /**
-     * Unit tests covering the DAO.allCitiesIn method
-     */
-    // test that cities list is empty when areaFilter and areaName are null
-    @Test
-    public void allCitiesIn_bothArgumentsNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-        String areaName = null;
-
-        // when
-        ArrayList<Record> cities = dao.allCitiesIn(areaFilter, areaName);
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // test that cities list is empty when areaFilter is null
-    @Test
-    public void allCitiesIn_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> cities = dao.allCitiesIn(areaFilter, "Scotland");
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // test that cities list is empty when areaName is null
-    @Test
-    public void allCitiesIn_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> cities = dao.allCitiesIn(App.COUNTRY, areaName);
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // check cities list is empty if called with an invalid areaFilter
-    @Test
-    public void allCitiesIn_areaFilterWrongListEmpty()
-    {
-        // given
-        String areaFilter = App.CITY;
-
-        // when
-        ArrayList<Record> cities = dao.allCitiesIn(areaFilter, "Glasgow");
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void allCitiesIn_happyPath()
-    {
-        // given
-        String areaFilter = App.COUNTRY;
-        String areaName = "France";
-
-        // then
-        dao.allCitiesIn(areaFilter, areaName); // No Error
-    }
-
-    /**
-     * Unit tests covering the DAO.topNCitiesIn method
-     */
-    // test that cities list is empty when areaFilter and areaName are null
-    @Test
-    public void topNCitiesIn_bothArgumentsNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-        String areaName = null;
-
-        // when
-        ArrayList<Record> cities = dao.topNCitiesIn(areaFilter, areaName, 5);
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // test that cities list is empty when areaFilter is null
-    @Test
-    public void topNCitiesIn_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> cities = dao.topNCitiesIn(areaFilter, "Europe", 10);
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // test that cities list is empty when areaName is null
-    @Test
-    public void topNCitiesIn_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> cities = dao.topNCitiesIn(App.COUNTRY, areaName, 3);
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // test that n cannot be less than or equal to 0
-    @Test
-    public void topNCitiesIn_negativeNListEmpty()
-    {
-        // given
-        Integer n = -3;
-
-        // when
-        ArrayList<Record> cities = dao.topNCitiesIn(App.COUNTRY, "China", n);
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // check cities list is empty if called with an invalid areaFilter
-    @Test
-    public void topNCitiesIn_areaFilterWrongListEmpty() {
-        // given
-        String areaFilter = App.CITY;
-
-        // when
-        ArrayList<Record> cities = dao.topNCitiesIn(areaFilter, "Houston", 15);
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void topNCitiesIn_happyPath()
-    {
-        // given
-        String areaFilter = App.WORLD;
-        String areaName = "";
-
-        // then
-        dao.topNCitiesIn(areaFilter, areaName, 15); // No Error
-    }
-
-    /**
-     * Unit tests covering the DAO.allCapitalCitiesIn method
-     */
-    // test that capitalCities list is empty when areaFilter and areaName are null
-    @Test
-    public void allCapitalCitiesIn_bothArgumentsNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-        String areaName = null;
-
-        // when
-        ArrayList<Record> capitalCities = dao.allCapitalCitiesIn(areaFilter, areaName);
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // test that capitalCities list is empty when areaFilter is null
-    @Test
-    public void allCapitalCitiesIn_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> capitalCities = dao.allCapitalCitiesIn(areaFilter, "Scotland");
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // test that capitalCities list is empty when areaName is null
-    @Test
-    public void allCapitalCitiesIn_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> capitalCities = dao.allCapitalCitiesIn(App.COUNTRY, areaName);
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // check capital cities list is empty if called with an invalid areaFilter
-    @Test
-    public void allCapitalCitiesIn_areaFilterWrongListEmpty()
-    {
-        // given
-        String areaFilter = App.DISTRICT;
-
-        // when
-        ArrayList<Record> capitalCities = dao.allCapitalCitiesIn(areaFilter, "Glasgow");
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void allCapitalCitiesIn_happyPath()
-    {
-        // given
-        String areaFilter = App.COUNTRY;
-        String areaName = "France";
-
-        // then
-        dao.allCapitalCitiesIn(areaFilter, areaName); // No Error
-    }
-
-    /**
-     * Unit tests covering the DAO.topNCapitalCitiesIn method
-     */
-    // test that capitalCities list is empty when areaFilter and areaName are null
-    @Test
-    public void topNCapitalCitiesIn_bothArgumentsNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-        String areaName = null;
-
-        // when
-        ArrayList<Record> capitalCities = dao.topNCapitalCitiesIn(areaFilter, areaName, 5);
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // test that capitalCities list is empty when areaFilter is null
-    @Test
-    public void topNCapitalCitiesIn_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> capitalCities = dao.topNCapitalCitiesIn(areaFilter, "Europe", 10);
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // test that capitalCities list is empty when areaName is null
-    @Test
-    public void topNCapitalCitiesIn_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> capitalCities = dao.topNCapitalCitiesIn(App.COUNTRY, areaName, 3);
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // test that n cannot be less than or equal to 0
-    @Test
-    public void topNCapitalCitiesIn_negativeNListEmpty()
-    {
-        // given
-        Integer n = -3;
-
-        // when
-        ArrayList<Record> capitalCities = dao.topNCapitalCitiesIn(App.COUNTRY, "China", n);
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // check capital cities list is empty if called with an invalid areaFilter
-    @Test
-    public void topNCapitalCitiesIn_areaFilterWrongListEmpty() {
-        // given
-        String areaFilter = App.CITY;
-
-        // when
-        ArrayList<Record> capitalCities = dao.topNCapitalCitiesIn(areaFilter, "Canberra", 15);
-
-        // then
-        assertEquals(0, capitalCities.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void topNCapitalCitiesIn_happyPath()
-    {
-        // given
-        String areaFilter = App.REGION;
-        String areaName = "Nordic Countries";
-
-        // then
-        dao.topNCapitalCitiesIn(areaFilter, areaName, 15); // No Error
-    }
-
-    /**
-     * Unit tests covering the DAO.populationLivingInAndNotInCities method
-     */
-    // test that residenceReport list is empty when areaFilter and areaName are null
-    @Test
-    public void populationLivingInAndNotInCities_bothArgumentsNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-        String areaName = null;
-
-        // when
-        ArrayList<Record> residenceReport = dao.populationLivingInAndNotInCities(areaFilter, areaName);
-
-        // then
-        assertEquals(0, residenceReport.size());
-    }
-
-    // test that residenceReport list is empty when areaFilter is null
-    @Test
-    public void populationLivingInAndNotInCities_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> residenceReport = dao.populationLivingInAndNotInCities(areaFilter, "Scotland");
-
-        // then
-        assertEquals(0, residenceReport.size());
-    }
-
-    // test that residenceReport list is empty when areaName is null
-    @Test
-    public void populationLivingInAndNotInCities_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> residenceReport = dao.populationLivingInAndNotInCities(App.COUNTRY, areaName);
-
-        // then
-        assertEquals(0, residenceReport.size());
-    }
-
-    // check cities list is empty if called with an invalid areaFilter
-    @Test
-    public void populationLivingInAndNotInCities_areaFilterWrongListEmpty() {
-        // given
-        String areaFilter = App.WORLD;
-
-        // when
-        ArrayList<Record> cities = dao.populationLivingInAndNotInCities(areaFilter, "");
-
-        // then
-        assertEquals(0, cities.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void populationLivingInAndNotInCities_happyPath()
-    {
-        // given
-        String areaFilter = App.CONTINENT;
-        String areaName = "Asia";
-
-        // then
-        dao.populationLivingInAndNotInCities(areaFilter, areaName); // No Error
-    }
-
-     /**
-     * Unit tests covering the DAO.populationOf method
-     */
-    // test that population list is empty when areaFilter and areaName are null
-    @Test
-    public void populationOf_bothArgumentsNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-        String areaName = null;
-
-        // when
-        ArrayList<Record> population = dao.populationOf(areaFilter, areaName);
-
-        // then
-        assertEquals(0, population.size());
-    }
-
-    // test that population list is empty when areaFilter is null
-    @Test
-    public void populationOf_areaFilterNullListEmpty()
-    {
-        // given
-        String areaFilter = null;
-
-        // when
-        ArrayList<Record> population = dao.populationOf(areaFilter, "Scotland");
-
-        // then
-        assertEquals(0, population.size());
-    }
-
-    // test that population list is empty when areaName is null
-    @Test
-    public void populationOf_areaNameNullListEmpty()
-    {
-        // given
-        String areaName = null;
-
-        // when
-        ArrayList<Record> population = dao.populationOf(App.COUNTRY, areaName);
-
-        // then
-        assertEquals(0, population.size());
-    }
-
-    // check test passes with valid areaFilter and areaName
-    @Test
-    public void populationOf_happyPath()
-    {
-        // given
-        String areaFilter = App.CITY;
-        String areaName = "Glasgow";
-
-        // then
-        dao.populationOf(areaFilter, areaName); // No Error
-    }
-
-    /**
-     * Unit tests covering the DAO.languageReport method
-     */
-    // check no error is thrown when method is called with a null connection
-    @Test
-    public void languageReport_happyPath()
-    {
-        // then
-        dao.languageReport(); // No Error
-    }
-
-}
